[feat] 바텀시트 오픈시 오버레이 구간 변경 및 상단 바 색상 바뀜 구현
