+++ lib/feature/field/data/stadium_code.dart
@@ -1,0 +1,11 @@
+final Map<String, String> stadiumNameToCode = {
+  '잠실 야구장': 'JAM',
+  '고척 스카이돔': 'GOC',
+  '랜더스 필드': 'ICN',
+  '위즈 파크': 'SUW',
+  '한화생명 볼파크': 'DJN',
+  '라이온즈 파크': 'DAE',
+  '사직 야구장': 'BUS',
+  'NC 파크장': 'CHW',
+  '챔피언스 필드': 'GWJ',
+};
